Extract game board to component
