+++ src/Game.vue
@@ -1,358 +1,17 @@
 <script setup lang="ts">
-import { onUnmounted } from "vue"
 import { getWordOfTheDay, wordInDictionary } from "./words"
-import Keyboard from "./Keyboard.vue"
-import { LetterState } from "./types"
-import Message from "./Message.vue"
+import Maker from "./Maker.vue"
+import Board from "./Board.vue"
 
-const { word, sender } = getWordOfTheDay()
-let answer: string
-if (word == null) answer = ""
-
-// Board state. Each tile is represented as { letter, state }
-const board = $ref(
-  Array.from({ length: 6 }, () =>
-    Array.from({ length: 5 }, () => ({
-      letter: "",
-      state: LetterState.INITIAL,
-    }))
-  )
-)
-
-// Current active row.
-let currentRowIndex = $ref(0)
-const currentRow = $computed(() => board[currentRowIndex])
-
-// Feedback state: message and shake
-let message = $ref("")
-let grid = $ref("")
-let shakeRowIndex = $ref(-1)
-let success = $ref(false)
-
-// Keep track of revealed letters for the virtual keyboard
-const letterStates: Record<string, LetterState> = $ref({})
-
-// Handle keyboard input.
-let allowInput = true
-
-const onKeyup = (e: KeyboardEvent) => onKey(e.key)
-
-window.addEventListener("keyup", onKeyup)
-
-onUnmounted(() => {
-  window.removeEventListener("keyup", onKeyup)
-})
-
-function onKey(key: string) {
-  if (!allowInput) return
-  if (/^[a-zA-Z]$/.test(key)) {
-    fillTile(key.toLowerCase())
-  } else if (key === "Backspace") {
-    clearTile()
-  } else if (key === "Enter") {
-    completeRow()
-  }
-}
-
-function fillTile(letter: string) {
-  for (const tile of currentRow) {
-    if (!tile.letter) {
-      tile.letter = letter
-      break
-    }
-  }
-}
-
-function clearTile() {
-  for (const tile of [...currentRow].reverse()) {
-    if (tile.letter) {
-      tile.letter = ""
-      break
-    }
-  }
-}
-
-function completeRow() {
-  if (currentRow.every((tile) => tile.letter)) {
-    const guess = currentRow.map((tile) => tile.letter).join("")
-    if (
-      // If the guess is the answer, it must be allowed
-      guess !== answer &&
-      // Disallow a word that's not in the dictionary
-      wordInDictionary(guess) === "unknown" &&
-      // ...unless the answer is also not in the dictionary
-      wordInDictionary(answer) !== "unknown"
-    ) {
-      shake()
-      showMessage("Word not in dictionary")
-      return
-    }
-
-    const answerLetters: (string | null)[] = answer.split("")
-    // first pass: mark correct ones
-    currentRow.forEach((tile, i) => {
-      if (answerLetters[i] === tile.letter) {
-        tile.state = letterStates[tile.letter] = LetterState.CORRECT
-        answerLetters[i] = null
-      }
-    })
-    // second pass: mark the present
-    currentRow.forEach((tile) => {
-      if (!tile.state && answerLetters.includes(tile.letter)) {
-        tile.state = LetterState.PRESENT
-        answerLetters[answerLetters.indexOf(tile.letter)] = null
-        if (!letterStates[tile.letter]) {
-          letterStates[tile.letter] = LetterState.PRESENT
-        }
-      }
-    })
-    // 3rd pass: mark absent
-    currentRow.forEach((tile) => {
-      if (!tile.state) {
-        tile.state = LetterState.ABSENT
-        if (!letterStates[tile.letter]) {
-          letterStates[tile.letter] = LetterState.ABSENT
-        }
-      }
-    })
-
-    allowInput = false
-    if (currentRow.every((tile) => tile.state === LetterState.CORRECT)) {
-      // yay!
-      setTimeout(() => {
-        grid = genResultGrid()
-        showMessage(
-          ["Genius", "Magnificent", "Impressive", "Splendid", "Great", "Phew"][
-            currentRowIndex
-          ],
-          -1
-        )
-        success = true
-      }, 1600)
-    } else if (currentRowIndex < board.length - 1) {
-      // go the next row
-      currentRowIndex++
-      setTimeout(() => {
-        allowInput = true
-      }, 1600)
-    } else {
-      // game over :(
-      setTimeout(() => {
-        showMessage(answer.toUpperCase(), -1)
-      }, 1600)
-    }
-  } else {
-    shake()
-    showMessage("Not enough letters")
-  }
-}
-
-function showMessage(msg: string, time = 1000) {
-  message = msg
-  if (time > 0) {
-    setTimeout(() => {
-      message = ""
-    }, time)
-  }
-}
-
-function shake() {
-  shakeRowIndex = currentRowIndex
-  setTimeout(() => {
-    shakeRowIndex = -1
-  }, 1000)
-}
-
-const icons = {
-  [LetterState.CORRECT]: "🟩",
-  [LetterState.PRESENT]: "🟨",
-  [LetterState.ABSENT]: "⬜",
-  [LetterState.INITIAL]: null,
-}
-
-function genResultGrid() {
-  return board
-    .slice(0, currentRowIndex + 1)
-    .map((row) => {
-      return row.map((tile) => icons[tile.state]).join("")
-    })
-    .join("\n")
-}
+const { word } = getWordOfTheDay()
 </script>
 
 <template>
-  <Message :message="message"></Message>
   <header>
     <h1>CURDLE</h1>
     <h2><span>cu</span>stom wo<span>rdle</span></h2>
+    <a v-if="word" id="source-link" href="/">New</a>
   </header>
-  <div id="board">
-    <div
-      v-for="(row, index) in board"
-      :class="[
-        'row',
-        shakeRowIndex === index && 'shake',
-        success && currentRowIndex === index && 'jump',
-      ]"
-    >
-      <div
-        v-for="(tile, index) in row"
-        :class="['tile', tile.letter && 'filled', tile.state && 'revealed']"
-      >
-        <div class="front" :style="{ transitionDelay: `${index * 300}ms` }">
-          {{ tile.letter }}
-        </div>
-        <div
-          :class="['back', tile.state]"
-          :style="{
-            transitionDelay: `${index * 300}ms`,
-            animationDelay: `${index * 100}ms`,
-          }"
-        >
-          {{ tile.letter }}
-        </div>
-      </div>
-    </div>
-  </div>
-  <Keyboard @key="onKey" :letter-states="letterStates" />
+  <Board v-if="word" :answer="word"></Board>
+  <Maker v-else></Maker>
 </template>
-
-<style scoped>
-#board {
-  display: grid;
-  grid-template-rows: repeat(6, 1fr);
-  grid-gap: 5px;
-  padding: 10px;
-  box-sizing: border-box;
-  --height: min(420px, calc(var(--vh, 100vh) - 310px));
-  height: var(--height);
-  width: min(350px, calc(var(--height) / 6 * 5));
-  margin: 0px auto;
-}
-
-.row {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  grid-gap: 5px;
-}
-.tile {
-  width: 100%;
-  font-size: 2rem;
-  line-height: 2rem;
-  font-weight: bold;
-  vertical-align: middle;
-  text-transform: uppercase;
-  user-select: none;
-  position: relative;
-}
-.tile.filled {
-  animation: zoom 0.2s;
-}
-.tile .front,
-.tile .back {
-  box-sizing: border-box;
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  transition: transform 0.6s;
-  backface-visibility: hidden;
-  -webkit-backface-visibility: hidden;
-}
-.tile .front {
-  border: 2px solid #d3d6da;
-}
-.tile.filled .front {
-  border-color: #999;
-}
-.tile .back {
-  transform: rotateX(180deg);
-}
-.tile.revealed .front {
-  transform: rotateX(180deg);
-}
-.tile.revealed .back {
-  transform: rotateX(0deg);
-}
-
-@keyframes zoom {
-  0% {
-    transform: scale(1.1);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-
-.shake {
-  animation: shake 0.5s;
-}
-
-@keyframes shake {
-  0% {
-    transform: translate(1px);
-  }
-  10% {
-    transform: translate(-2px);
-  }
-  20% {
-    transform: translate(2px);
-  }
-  30% {
-    transform: translate(-2px);
-  }
-  40% {
-    transform: translate(2px);
-  }
-  50% {
-    transform: translate(-2px);
-  }
-  60% {
-    transform: translate(2px);
-  }
-  70% {
-    transform: translate(-2px);
-  }
-  80% {
-    transform: translate(2px);
-  }
-  90% {
-    transform: translate(-2px);
-  }
-  100% {
-    transform: translate(1px);
-  }
-}
-
-.jump .tile .back {
-  animation: jump 0.5s;
-}
-
-@keyframes jump {
-  0% {
-    transform: translateY(0px);
-  }
-  20% {
-    transform: translateY(5px);
-  }
-  60% {
-    transform: translateY(-25px);
-  }
-  90% {
-    transform: translateY(3px);
-  }
-  100% {
-    transform: translateY(0px);
-  }
-}
-
-@media (max-height: 680px) {
-  .tile {
-    font-size: 3vh;
-  }
-}
-</style>
